[AZTH] fixed smartstone function definition

+++ modules/azerothshard/src/plugins/AzthSmartStone.h
@@ -54,7 +54,7 @@ public:
     std::vector<SmartStoneCommand> ssCommands2;
 
     void loadCommands();
-    void SmartStoneSendListInventory(WorldSession * session, int vendorGuid);
+    void SmartStoneSendListInventory(WorldSession * session, uint64 vendorGuid);
     SmartStoneCommand getCommandById(uint32 id);
     SmartStoneCommand getCommandByItem(uint32 item);
     bool isNullCommand(SmartStoneCommand command);
